allow FileIdentifier to be resolved to a FileLocator
FileLocator has the url for the file, normally taken from the Location header when a file is, e.g., created. However, functions like read_document_files don't yield a FileLocator, as there is no Location header, so it can only return a FileIdentifier. In order for these FileIdentifiers to be used in a context where the file's location is needed, resolve_file_location can be used.